--- src/main/java/com/kh/saintra/anonymous/board/AnonymousDto.java
@@ -1,31 +1,0 @@
-package com.kh.saintra.anonymous.board;
-
-import java.util.Date;
-
-public class AnonymousDto {
-    private Long id;
-    private Long userId;
-    private String title;
-    private String content;
-    private Date createDate;
-    private String isActive;
-
-    // Getter/Setter
-    public Long getId() { return id; }
-    public void setId(Long id) { this.id = id; }
-
-    public Long getUserId() { return userId; }
-    public void setUserId(Long userId) { this.userId = userId; }
-
-    public String getTitle() { return title; }
-    public void setTitle(String title) { this.title = title; }
-
-    public String getContent() { return content; }
-    public void setContent(String content) { this.content = content; }
-
-    public Date getCreateDate() { return createDate; }
-    public void setCreateDate(Date createDate) { this.createDate = createDate; }
-
-    public String getIsActive() { return isActive; }
-    public void setIsActive(String isActive) { this.isActive = isActive; }
-}
