--- src/main/java/com/kh/saintra/anonymous/comment/CommentDto.java
@@ -1,31 +1,0 @@
-package com.kh.saintra.anonymous.comment;
-
-import java.util.Date;
-
-public class CommentDto {
-    private Long id;
-    private Long boardId;
-    private Long userId;
-    private String content;
-    private Date createDate;
-    private String isActive;
-
-    // Getters & Setters
-    public Long getId() { return id; }
-    public void setId(Long id) { this.id = id; }
-
-    public Long getBoardId() { return boardId; }
-    public void setBoardId(Long boardId) { this.boardId = boardId; }
-
-    public Long getUserId() { return userId; }
-    public void setUserId(Long userId) { this.userId = userId; }
-
-    public String getContent() { return content; }
-    public void setContent(String content) { this.content = content; }
-
-    public Date getCreateDate() { return createDate; }
-    public void setCreateDate(Date createDate) { this.createDate = createDate; }
-
-    public String getIsActive() { return isActive; }
-    public void setIsActive(String isActive) { this.isActive = isActive; }
-}
