+++ src/main/java/com/kh/saintra/anonymous/model/dto/CommentDto.java
@@ -1,0 +1,15 @@
+package com.kh.saintra.anonymous.model.dto;
+
+import lombok.Data;
+import java.util.Date;
+
+@Data
+public class CommentDto {
+    private Long id;
+    private Long boardId;
+    private Long userId;
+    private String content;
+    private Date createDate;
+    private String isActive;
+}
+
